fix: only display ai bubble button on home page
