+++ frontend/appflowy_flutter/lib/mobile/presentation/home/mobile_home_page.dart
@@ -1,15 +1,11 @@
 import 'dart:io';
 
-import 'package:appflowy/generated/flowy_svgs.g.dart';
-import 'package:appflowy/generated/locale_keys.g.dart';
-import 'package:appflowy/mobile/presentation/base/gesture.dart';
 import 'package:appflowy/mobile/presentation/home/mobile_home_page_header.dart';
 import 'package:appflowy/mobile/presentation/home/tab/mobile_space_tab.dart';
 import 'package:appflowy/mobile/presentation/home/tab/space_order_bloc.dart';
 import 'package:appflowy/startup/startup.dart';
 import 'package:appflowy/user/application/auth/auth_service.dart';
 import 'package:appflowy/user/application/reminder/reminder_bloc.dart';
-import 'package:appflowy/util/theme_extension.dart';
 import 'package:appflowy/workspace/application/favorite/favorite_bloc.dart';
 import 'package:appflowy/workspace/application/menu/sidebar_sections_bloc.dart';
 import 'package:appflowy/workspace/application/recent/cached_recent_service.dart';
@@ -23,8 +19,6 @@ import 'package:appflowy_backend/protobuf/flowy-folder/view.pb.dart';
 import 'package:appflowy_backend/protobuf/flowy-folder/workspace.pb.dart';
 import 'package:appflowy_backend/protobuf/flowy-user/protobuf.dart';
 import 'package:appflowy_editor/appflowy_editor.dart';
-import 'package:easy_localization/easy_localization.dart';
-import 'package:flowy_infra_ui/flowy_infra_ui.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:provider/provider.dart';
@@ -140,19 +134,7 @@ class _MobileHomePageState extends State<MobileHomePage> {
           value: getIt<ReminderBloc>()..add(const ReminderEvent.started()),
         ),
       ],
-      child: Stack(
-        children: [
-          _HomePage(userProfile: widget.userProfile),
-          // only show ai chat button for cloud user
-          if (widget.userProfile.authenticator == AuthenticatorPB.AppFlowyCloud)
-            Positioned(
-              bottom: MediaQuery.of(context).padding.bottom + 16,
-              left: 20,
-              right: 20,
-              child: const _FloatingAIEntry(),
-            ),
-        ],
-      ),
+      child: _HomePage(userProfile: widget.userProfile),
     );
   }
 
@@ -240,80 +222,6 @@ class _HomePage extends StatelessWidget {
           ],
         );
       },
-    );
-  }
-}
-
-// disable ai chat for local model user
-class _FloatingAIEntry extends StatelessWidget {
-  const _FloatingAIEntry();
-
-  @override
-  Widget build(BuildContext context) {
-    return AnimatedGestureDetector(
-      scaleFactor: 0.99,
-      onTapUp: () => mobileCreateNewAIChatNotifier.value =
-          mobileCreateNewAIChatNotifier.value + 1,
-      child: DecoratedBox(
-        decoration: _buildShadowDecoration(context),
-        child: Container(
-          decoration: _buildWrapperDecoration(context),
-          height: 48,
-          alignment: Alignment.centerLeft,
-          child: Padding(
-            padding: const EdgeInsets.only(left: 18),
-            child: _buildHintText(context),
-          ),
-        ),
-      ),
-    );
-  }
-
-  BoxDecoration _buildShadowDecoration(BuildContext context) {
-    return BoxDecoration(
-      borderRadius: BorderRadius.circular(30),
-      boxShadow: [
-        BoxShadow(
-          blurRadius: 20,
-          spreadRadius: 1,
-          offset: const Offset(0, 4),
-          color: Colors.black.withOpacity(0.05),
-        ),
-      ],
-    );
-  }
-
-  BoxDecoration _buildWrapperDecoration(BuildContext context) {
-    final outlineColor = Theme.of(context).colorScheme.outline;
-    final borderColor = Theme.of(context).isLightMode
-        ? outlineColor.withOpacity(0.7)
-        : outlineColor.withOpacity(0.3);
-    return BoxDecoration(
-      borderRadius: BorderRadius.circular(30),
-      color: Theme.of(context).colorScheme.surface,
-      border: Border.fromBorderSide(
-        BorderSide(
-          color: borderColor,
-        ),
-      ),
-    );
-  }
-
-  Widget _buildHintText(BuildContext context) {
-    return Row(
-      children: [
-        FlowySvg(
-          FlowySvgs.toolbar_item_ai_s,
-          size: const Size.square(16.0),
-          color: Theme.of(context).hintColor,
-          opacity: 0.7,
-        ),
-        const HSpace(8),
-        FlowyText(
-          LocaleKeys.chat_inputMessageHint.tr(),
-          color: Theme.of(context).hintColor,
-        ),
-      ],
     );
   }
 }
